회원가입 부분 코드 수정
