--- src/components/Signup/testData.json
@@ -1,33 +1,0 @@
-{
-  "httpStatus": 200,
-  "message": "조회 성공",
-  "results": {
-    "userId": 2020100000,
-    "subject": [
-      {
-        "subjectId": 1234,
-        "completeYear": "2022",
-        "completeTerm": "2학기 정규",
-        "subjectName": "데이터통신",
-        "subjectProfessor": "조충희",
-        "subjectSort": "전공선택"
-      },
-      {
-        "subjectId": 5678,
-        "completeYear": "2023",
-        "completeTerm": "1학기 정규",
-        "subjectName": "데이터베이스",
-        "subjectProfessor": "김영희",
-        "subjectSort": "전공필수"
-      },
-      {
-        "subjectId": 9012,
-        "completeYear": "2023",
-        "completeTerm": "2학기 정규",
-        "subjectName": "인공지능",
-        "subjectProfessor": "김현규",
-        "subjectSort": "전공필수"
-      }
-    ]
-  }
-}
